+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Clarify that some settings only apply to the full viewer
+- Add a new global setting and its site-specific counterpart for showing the
+  media selector page block only if there are several media
+- Fix "Show download link" setting
+
 ## [0.9.0] - 2025-10-28
 
 - Split the viewer into multiple components that can be used separately